CMake: added option to compile with debug flags. Changed enum to enum class. Almost implemented deserialization

+++ General/include/Grammar/CRule.hpp
@@ -6,7 +6,7 @@
 
 namespace formals { namespace grammars {
 
-    using ruleSymbolValyeType = uint32_t;
+    using ruleSymbolValyeType = size_t;
 
     struct RuleSymbol {
         ruleSymbolValyeType value;
